Create recommendation-based-on-user-input.js
This program prompts the user for input, and makes a selection from the object based on that input, using an if statement. It then displays the selection on the web page. This specific application recommends an outdoor activity based on the weather input from the user. This program could be modified to any scenario where specific criteria are tied to an object element, such as comparing how much money a user has (user input) to what they can purchase at the store (object elements).