Create hangman-game.js
This is a basic Hangman game. The program accepts user input for both the word to guess and guessed letters. As correct letters are guess, they are filled in in the correct guesses section. As incorrect letters are guess, the user accumulates a letter (the letters of HANGMAN) for each incorrect guess, until they have spelled hangman. For HTML see hangman-game.html.